Code has been refactored to reduce redundancy; Service and capabilityDelegation field has been implemented
