+++ x/ssi/keeper/did.go
@@ -11,13 +11,13 @@ import (
 )
 
 func (k Keeper) GetDidCount(ctx sdk.Context) uint64 {
-	// Get the store using storeKey (which is "blog") and PostCountKey (which is "Post-count-")
+	// Get the store using storeKey and DidCountKey (which is "Did-count-")
 	store := prefix.NewStore(ctx.KVStore(k.storeKey), []byte(types.DidCountKey))
-	// Convert the PostCountKey to bytes
+	// Convert the DidCountKey to bytes
 	byteKey := []byte(types.DidCountKey)
 	// Get the value of the count
 	bz := store.Get(byteKey)
-	// Return zero if the count value is not found (for example, it's the first post)
+	// Return zero if the count value is not found
 	if bz == nil {
 		return 0
 	}
@@ -32,16 +32,14 @@ func (k Keeper) HasDid(ctx sdk.Context, id string) bool {
 }
 
 // Retrieves the DID from the store
-func (k Keeper) GetDid(ctx *sdk.Context, id string) (*types.DidDocStructCreateDID, error) {
+func (k Keeper) GetDid(ctx *sdk.Context, id string) (*types.Did, error) {
 	store := prefix.NewStore(ctx.KVStore(k.storeKey), types.KeyPrefix(types.DidKey))
 
 	if !k.HasDid(*ctx, id) {
 		return nil, sdkerrors.ErrNotFound
 	}
 
-	// TODO: Currently, we can interchange DidDocStructCreateDID with DidDocStructUpdateDID
-	// as they are similar. They need to have a similar Interface
-	var value types.DidDocStructCreateDID
+	var value types.Did
 	var bytes = store.Get([]byte(id))
 	if err := k.cdc.Unmarshal(bytes, &value); err != nil {
 		return nil, sdkerrors.Wrap(sdkerrors.ErrInvalidType, err.Error())
@@ -51,14 +49,14 @@ func (k Keeper) GetDid(ctx *sdk.Context, id string) (*types.DidDocStructCreateDI
 }
 
 func (k Keeper) SetDidCount(ctx sdk.Context, count uint64) {
-	// Get the store using storeKey (which is "blog") and PostCountKey (which is "Post-count-")
+	// Get the store using storeKey and SchemaCountKey (which is "Did-count-")
 	store := prefix.NewStore(ctx.KVStore(k.storeKey), []byte(types.DidCountKey))
-	// Convert the PostCountKey to bytes
+	// Convert the DidCountKey to bytes
 	byteKey := []byte(types.DidCountKey)
 	// Convert count from uint64 to string and get bytes
 	bz := make([]byte, 8)
 	binary.BigEndian.PutUint64(bz, count)
-	// Set the value of Post-count- to count
+	// Set the value of Did-count- to count
 	store.Set(byteKey, bz)
 }
 
@@ -71,15 +69,14 @@ func (k Keeper) SetDid(ctx sdk.Context, did types.Did) error {
 }
 
 func (k Keeper) AppendDID(ctx sdk.Context, didSpec types.Did) uint64 {
-	// Get the current number of posts in the store
+	// Get the current number of DIDs in the store
 	count := k.GetDidCount(ctx)
 	// Get the store
 	store := prefix.NewStore(ctx.KVStore(k.storeKey), []byte(types.DidKey))
-	// Convert the post ID into bytes
-	// Marshal the post into bytes
+	// Marshal the DID into bytes
 	didDocString := k.cdc.MustMarshal(&didSpec)
 	store.Set(utils.UnsafeStrToBytes(didSpec.Id), didDocString)
-	// Update the post count
+	// Update the Did count
 	k.SetDidCount(ctx, count+1)
 	return count
 }
