+++ x/ssi/keeper/grpc_query_did.go
@@ -20,3 +20,20 @@ func (k Keeper) DidDocCount(goCtx context.Context, req *types.QueryDidDocCountRe
 
 	return &types.QueryDidDocCountResponse{Count: didDocCount}, nil
 }
+
+func (k Keeper) GetDidDocById(goCtx context.Context, req *types.QueryGetDidDocByIdRequest) (*types.QueryGetDidDocByIdResponse, error) {
+	if req == nil {
+		return nil, status.Error(codes.InvalidArgument, "invalid request")
+	}
+
+	ctx := sdk.UnwrapSDKContext(goCtx)
+
+	didDoc, err := k.GetDid(&ctx, req.DidDocId)
+	if err != nil {
+		return nil, status.Error(codes.NotFound, "DidDoc not found")
+	}
+
+	return &types.QueryGetDidDocByIdResponse{
+		DidDoc: didDoc,
+	}, nil
+}
