+++ x/ssi/types/message_schema.go
