+++ x/ssi/utils/did_vars.go
@@ -1,0 +1,8 @@
+package utils
+
+var ServiceTypes = []string{
+	"LinkedDomains",
+	"DIDCommMessaging",
+}
+
+var DidPrefix string = "did:hs:"
